pyproject.toml
[tool.poetry]
name = "zoom-api-socio"
version = "0.1.0"
description = "API Zoom Para Projetos Sociais"
authors = ["Guilherme D.A"]
readme = "README.md"

[tool.taskipy.tasks]
lint = "blue ."

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.1.0"
psycopg2 = "^2.9.6"
django-filter = "^23.2"
blue = "^0.9.1"
isort = "^5.12.0"
django-split-settings = "^1.2.0"
hypercorn = "^0.14.3"
python-dotenv = "^1.0.0"
django-redis = "^5.3.0"
colorlog = "^6.7.0"
channels-redis = "^4.1.0"
drf-yasg = "^1.21.6"
djangorestframework-simplejwt = "^5.2.2"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.src.masonry.api"
